move scanner package_meta_result to util/type_util
