+++ depobs/clients/cratesio.py
@@ -9,7 +9,7 @@ from depobs.scanner.models.rust import (
     RustPackageID,
     cargo_metadata_to_rust_crates,
 )
-from depobs.scanner.models.package_meta_result import Result
+from depobs.util.type_util import Result
 
 
 log = logging.getLogger(__name__)
