+++ depobs/clients/npm_registry.py
@@ -6,7 +6,7 @@ from typing import Any, AsyncGenerator, Dict, Iterable, Optional, TypedDict
 import logging
 
 from depobs.clients.aiohttp_client_config import AIOHTTPClientConfig
-from depobs.scanner.models.package_meta_result import Result
+from depobs.util.type_util import Result
 from depobs.util.serialize_util import grouper
 from depobs.util.traceback_util import exc_to_str
