+++ depobs/util/type_util.py
